backend/requirements.txt
PyAudio
websockets
janus
fastapi
uvicorn[standard]
python-socketio
python-dotenv
requests
python-socketio[asyncio_client]
sentence-transformers
faiss-cpu
numpy
pymongo
google-api-python-client
google-auth-httplib2
google-auth-oauthlib